[PATRO23-43] move budgets classes to Shared folder
