+++ src/modules/budget/application/Budget/Shared/ImportingBudgets/GetImportResult.cs
@@ -1,6 +1,6 @@
 using Intive.Patronage2023.Modules.Budget.Application.Data;
 
-namespace Intive.Patronage2023.Modules.Budget.Application.Budget.ExportingBudgets;
+namespace Intive.Patronage2023.Modules.Budget.Application.Budget.Shared.ImportingBudgets;
 
 /// <summary>
 /// GetImportResult command.
