+++ src/modules/budget/application/Data/Service/BudgetImportService.cs
@@ -1,7 +1,7 @@
 using System.Globalization;
 using Azure.Storage.Blobs;
 using CsvHelper.Configuration;
-using Intive.Patronage2023.Modules.Budget.Application.Budget.ExportingBudgets;
+using Intive.Patronage2023.Modules.Budget.Application.Budget.Shared.ImportingBudgets;
 using Intive.Patronage2023.Modules.Budget.Domain;
 using Intive.Patronage2023.Shared.Abstractions;
 using Microsoft.AspNetCore.Http;
