+++ src/modules/budget/application/Data/Service/IBudgetImportService.cs
@@ -1,4 +1,4 @@
-using Intive.Patronage2023.Modules.Budget.Application.Budget.ExportingBudgets;
+using Intive.Patronage2023.Modules.Budget.Application.Budget.Shared.ImportingBudgets;
 using Microsoft.AspNetCore.Http;
 
 namespace Intive.Patronage2023.Modules.Budget.Application.Data.Service;
